# Replace numb-bigint / crypto-bigint with `rug` and lots of other small optimizations for bigint2