--- a/risc0/circuit/rv32im/src/prove/emu/addr.rs
+++ b/risc0/circuit/rv32im/src/prove/emu/addr.rs
@@ -21,7 +21,7 @@ use super::pager::PAGE_WORDS;
 #[derive(Clone, Copy, Eq, Hash, PartialEq)]
 pub struct ByteAddr(pub u32);
 
-#[derive(Clone, Copy, Eq, Hash, PartialEq)]
+#[derive(Clone, Copy, Eq, PartialOrd, Ord, Hash, PartialEq)]
 pub struct WordAddr(pub u32);
 
 impl From<ByteAddr> for WordAddr {