--- a/risc0/circuit/rv32im/src/prove/emu/bibc.rs
+++ b/risc0/circuit/rv32im/src/prove/emu/bibc.rs
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use core::ops::Shl;
 use std::io::Read;
 
 use anyhow::{anyhow, Result};
 use byteorder::{LittleEndian, ReadBytesExt};
-use num_bigint::{BigInt, BigUint, Sign};
 use num_derive::FromPrimitive;
 use num_traits::FromPrimitive;
+use rug::{Assign, Integer};
 
 #[derive(Debug)]
 pub(crate) struct Type {
@@ -65,9 +66,9 @@ pub(crate) struct Program {
 }
 
 pub(crate) trait BigIntIO {
-    fn load(&mut self, arena: u32, offset: u32, count: u32) -> Result<BigUint>;
+    fn load(&mut self, arena: u32, offset: u32, count: u32) -> Result<Integer>;
 
-    fn store(&mut self, arena: u32, offset: u32, count: u32, value: &BigUint) -> Result<()>;
+    fn store(&mut self, arena: u32, offset: u32, count: u32, value: &Integer) -> Result<()>;
 }
 
 impl Op {
@@ -151,65 +152,55 @@ impl Program {
     }
 
     pub fn eval<T: BigIntIO>(&self, io: &mut T) -> Result<()> {
-        let mut regs = vec![BigInt::ZERO; self.ops.len()];
+        let mut regs = vec![Integer::ZERO; self.ops.len()];
         for (op_index, op) in self.ops.iter().enumerate() {
             tracing::debug!("[{op_index}]: {op:?}");
             match op.code {
                 OpCode::Const => {
                     let offset = op.a;
                     let words = op.b;
-                    let mut value = BigUint::from(0_u64);
+                    let mut value = Integer::from(0_u64);
                     for i in 0..words {
                         let word: u64 = self.constants[offset + i];
-                        let mut tmp = BigUint::from(word);
-                        tmp <<= i * 64;
-                        value |= &tmp;
+                        value |= word.shl(i * 64);
                     }
-                    regs[op_index] = BigInt::from_biguint(Sign::Plus, value);
+                    regs[op_index] = value;
                 }
                 OpCode::Load => {
                     let typ = &self.types[op.result_type];
                     let count = typ.coeffs.next_multiple_of(16) as u32;
                     let value = io.load(op.arena(), op.offset(), count)?;
-                    regs[op_index] = BigInt::from_biguint(Sign::Plus, value);
+                    regs[op_index] = value;
                 }
                 OpCode::Store => {
                     let typ = &self.types[op.result_type];
                     let count = typ.coeffs.next_multiple_of(16) as u32;
                     let value = &regs[op.b];
-                    let value = value.to_biguint().ok_or_else(|| {
-                        anyhow!("Negative output produced during bigint2 acceleration")
-                    })?;
-                    io.store(op.arena(), op.offset(), count, &value)?;
+                    io.store(op.arena(), op.offset(), count, value)?;
                 }
                 OpCode::Add => {
-                    let (lhs, rhs) = operands(op, op_index, &regs);
-                    regs[op_index] = lhs + rhs;
+                    let (lhs, rhs, dst) = operands_mut(op, op_index, &mut regs);
+                    dst.assign(lhs + rhs);
                 }
                 OpCode::Sub => {
-                    let (lhs, rhs) = operands(op, op_index, &regs);
-                    regs[op_index] = lhs - rhs;
+                    let (lhs, rhs, dst) = operands_mut(op, op_index, &mut regs);
+                    dst.assign(lhs - rhs);
                 }
                 OpCode::Mul => {
-                    let (lhs, rhs) = operands(op, op_index, &regs);
-                    regs[op_index] = lhs * rhs;
+                    let (lhs, rhs, dst) = operands_mut(op, op_index, &mut regs);
+                    dst.assign(lhs * rhs);
                 }
                 OpCode::Rem => {
-                    let (lhs, rhs) = operands(op, op_index, &regs);
-                    let value = lhs % rhs;
-                    regs[op_index] = value;
+                    let (lhs, rhs, dst) = operands_mut(op, op_index, &mut regs);
+                    dst.assign(lhs % rhs);
                 }
                 OpCode::Quo => {
-                    let (lhs, rhs) = operands(op, op_index, &regs);
-                    let value = lhs / rhs;
-                    regs[op_index] = value;
+                    let (lhs, rhs, dst) = operands_mut(op, op_index, &mut regs);
+                    dst.assign(lhs / rhs);
                 }
                 OpCode::Inv => {
-                    let (lhs, rhs) = operands(op, op_index, &regs);
-                    let value = lhs
-                        .modinv(rhs)
-                        .ok_or_else(|| anyhow!("Can't divide by zero"))?;
-                    regs[op_index] = value;
+                    let (lhs, rhs, dst) = operands_mut(op, op_index, &mut regs);
+                    dst.assign(lhs.clone().invert(rhs).unwrap()); // TODO unwrap and clone
                 }
             }
         }
@@ -217,8 +208,35 @@ impl Program {
     }
 }
 
-fn operands<'p>(op: &Op, op_index: usize, regs: &'p [BigInt]) -> (&'p BigInt, &'p BigInt) {
+// fn operands<'p>(op: &Op, op_index: usize, regs: &'p [Integer]) -> (&'p Integer, &'p Integer) {
+//     assert!(op.a < op_index);
+//     assert!(op.b < op_index);
+//     (&regs[op.a], &regs[op.b])
+// }
+
+fn operands_mut<'a>(
+    op: &Op,
+    op_index: usize,
+    regs: &'a mut [Integer],
+) -> (&'a Integer, &'a Integer, &'a mut Integer) {
     assert!(op.a < op_index);
     assert!(op.b < op_index);
-    (&regs[op.a], &regs[op.b])
+    // op_index must be strictly greater than a and b.
+
+    // 1) Split the slice at `op_index`.
+    //    - `prefix` is `[0..op_index]` (indices 0 through op_index-1).
+    //    - `suffix` is `[op_index..]` (indices op_index through end).
+    let (prefix, suffix) = regs.split_at_mut(op_index);
+
+    // 2) `prefix` can be borrowed immutably for `lhs` and `rhs`.
+    let lhs = &prefix[op.a];
+    let rhs = &prefix[op.b];
+
+    // 3) `suffix` starts exactly at `op_index`, so `suffix[0]` is `regs[op_index]`.
+    let dst = &mut suffix[0];
+
+    // 4) Now we have disjoint references:
+    //    - `lhs`, `rhs` from the prefix (immutable)
+    //    - `dst` from the suffix (mutable)
+    (lhs, rhs, dst)
 }