--- a/risc0/circuit/rv32im/src/prove/emu/pager.rs
+++ b/risc0/circuit/rv32im/src/prove/emu/pager.rs
@@ -1,4 +1,4 @@
-// Copyright 2024 RISC Zero, Inc.
+// Copyright 2025 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.