--- a/risc0/circuit/rv32im/src/prove/emu/preflight/bigint2.rs
+++ b/risc0/circuit/rv32im/src/prove/emu/preflight/bigint2.rs
@@ -175,9 +175,7 @@ impl BytePolynomial {
 
     fn shift(&self) -> Self {
         let mut ret = self.coeffs.clone();
-        for _ in 0..BIGINT2_WIDTH_BYTES {
-            ret.insert(0, 0);
-        }
+        ret.resize(ret.len() + BIGINT2_WIDTH_BYTES, 0);
         Self { coeffs: ret }
     }