Sample bigint2 workload:
```
exec 3s -> 1s
preflight 500ms -> 291ms 
```

I left a few TODO's and a unwrap that need to be cleaned up, opening PR for visibility for @bobbobbio to help with and check CI.